+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [1.4.0] 2022-11-08
+
+### Added
+- behaviorDescription to components Addons
 - New title and paragraph elements to email editor
 
 ## [1.3.0] 2022-08-16